Image Similarity Cortex Search/requirements.txt
python-dotenv
openai
ipykernel
ipywidgets
matplotlib
diffusers
transformers
scipy
ftfy
accelerate
mlflow
tenacity
tqdm
faiss-cpu
bokeh
torch
torchvision
pillow
snowflake-snowpark-python
Exif
bokeh
imageio